+++ src/test/neural_network_test.rs
@@ -24,6 +24,7 @@ mod global_max_pooling_2d_test;
 mod global_max_pooling_3d_test;
 mod group_normalization_test;
 mod gru_test;
+mod guassian_noise_test;
 mod instance_normalization_test;
 mod layer_normalization_test;
 mod layer_weight_test;